requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11
httpx==0.27.0
tenacity==8.3.0

# NLP / Embeddings
sentence-transformers==3.0.0
torch==2.3.0
transformers==4.41.1

# Vector DBs
faiss-cpu==1.8.0
pinecone-client==3.2.2

# Data
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
